+++ renderer/components/Lnurl/index.js
@@ -1,2 +1,3 @@
 export LnurlAuthPrompt from './LnurlAuthPrompt'
 export LnurlChannelPrompt from './LnurlChannelPrompt'
+export LnurlWithdrawPrompt from './LnurlWithdrawPrompt'
