+++ renderer/components/Pay/index.js
@@ -10,4 +10,3 @@ export PayPanelHeader from './PayPanelHeader'
 export PaySummary from './PaySummary'
 export PaySummaryLightning from './PaySummaryLightning'
 export PaySummaryOnChain from './PaySummaryOnChain'
-export LnurlWithdrawPrompt from './LnurlWithdrawPrompt'
