+++ renderer/containers/App/App.js
@@ -7,8 +7,12 @@ import { updateAutopilotNodeScores } from 'reducers/autopilot'
 import { initActivityHistory } from 'reducers/activity'
 import { fetchTransactions } from 'reducers/transaction'
 import { appSelectors } from 'reducers/app'
-import { finishLnurlAuth, finishLnurlChannel, lnurlSelectors } from 'reducers/lnurl'
-import { finishLnurlWithdraw, paySelectors } from 'reducers/pay'
+import {
+  finishLnurlAuth,
+  finishLnurlChannel,
+  finishLnurlWithdraw,
+  lnurlSelectors,
+} from 'reducers/lnurl'
 import { initBackupService } from 'reducers/backup'
 import { setModals, modalSelectors } from 'reducers/modal'
 import { fetchSuggestedNodes } from 'reducers/channels'
@@ -21,9 +25,9 @@ const mapStateToProps = state => ({
   isAppReady: appSelectors.isAppReady(state),
   redirectPayReq: state.pay.redirectPayReq,
   modals: modalSelectors.getModalState(state),
-  lnurlWithdrawParams: paySelectors.lnurlWithdrawParams(state),
+  lnurlWithdrawParams: lnurlSelectors.lnurlWithdrawParams(state),
   willShowLnurlAuthPrompt: lnurlSelectors.willShowLnurlAuthPrompt(state),
-  willShowLnurlWithdrawPrompt: paySelectors.willShowLnurlWithdrawPrompt(state),
+  willShowLnurlWithdrawPrompt: lnurlSelectors.willShowLnurlWithdrawPrompt(state),
   willShowLnurlChannelPrompt: lnurlSelectors.willShowLnurlChannelPrompt(state),
 })
