+++ renderer/containers/Lnurl/LnurlWithdrawPrompt.js
@@ -1,14 +1,14 @@
 import { connect } from 'react-redux'
 import {
   finishLnurlWithdraw,
-  paySelectors,
   clearLnurlWithdraw,
   declineLnurlWithdraw,
-} from 'reducers/pay'
-import { LnurlWithdrawPrompt } from 'components/Pay'
+  lnurlSelectors,
+} from 'reducers/lnurl'
+import { LnurlWithdrawPrompt } from 'components/Lnurl'
 
 const mapStateToProps = state => ({
-  params: paySelectors.lnurlWithdrawParams(state),
+  params: lnurlSelectors.lnurlWithdrawParams(state),
 })
 
 const mapDispatchToProps = {
