+++ renderer/reducers/lnurl/constants.js
@@ -3,3 +3,6 @@ export const CLEAR_LNURL_AUTH = 'CLEAR_LNURL_AUTH'
 
 export const SET_LNURL_CHANNEL_PARAMS = 'SET_LNURL_CHANNEL_PARAMS'
 export const CLEAR_LNURL_CHANNEL = 'CLEAR_LNURL_CHANNEL'
+
+export const SET_LNURL_WITHDRAW_PARAMS = 'SET_LNURL_WITHDRAW_PARAMS'
+export const CLEAR_LNURL_WITHDRAW = 'CLEAR_LNURL_WITHDRAW'
