+++ renderer/reducers/lnurl/ipc.js
@@ -1,6 +1,6 @@
 import { getIntl } from '@zap/i18n'
 import { showNotification, showError, updateNotification } from 'reducers/notification'
-import { setLnurlAuthParams, setLnurlChannelParams } from './reducer'
+import { setLnurlAuthParams, setLnurlChannelParams, setLnurlWithdrawParams } from './reducer'
 import messages from './messages'
 
 // ------------------------------------
@@ -107,4 +107,68 @@ export const lnurlChannelError = (event, { service, reason }) => dispatch => {
       }
     )
   )
+}
+
+// ------------------------------------
+// lnurl-withdraw
+// ------------------------------------
+
+/**
+ * lnurlWithdrawSuccess - IPC handler for lnurlWithdrawSuccess event.
+ *
+ * @param {event} event Event ipc event
+ * @param {object} params { service }
+ * @param {string} params.service service
+ * @returns {(dispatch:Function) => void} Thunk
+ */
+export const lnurlWithdrawSuccess = (event, { service }) => dispatch => {
+  const intl = getIntl()
+  dispatch(
+    updateNotification(
+      { payload: { service } },
+      {
+        variant: 'success',
+        message: intl.formatMessage(messages.lnurl_withdraw_success, { service }),
+        timeout: 3000,
+        isProcessing: false,
+      }
+    )
+  )
+}
+
+/**
+ * lnurlWithdrawError - IPC handler for lnurlWithdrawError event.
+ *
+ * @param {event} event Event ipc event
+ * @param {object} params { service, reason }
+ * @param {string} params.service lnurl
+ * @param {string} params.reason error reason
+ * @returns {(dispatch:Function) => void} Thunk
+ */
+export const lnurlWithdrawError = (event, { service, reason }) => dispatch => {
+  const intl = getIntl()
+  dispatch(
+    updateNotification(
+      { payload: { service } },
+      {
+        variant: 'error',
+        message: intl.formatMessage(messages.lnurl_withdraw_error, { service, reason }),
+        isProcessing: false,
+      }
+    )
+  )
+}
+
+/**
+ * lnurlWithdrawRequest - IPC handler for lnurlWithdrawRequest event.
+ *
+ * @param {event} event Event ipc event
+ * @param {object} params { service, amount, memo }
+ * @param {string} params.service lnurl
+ * @param {number} params.amount ln pr amount
+ * @param {string} params.memo ln pr memo
+ * @returns {(dispatch:Function) => void} Thunk
+ */
+export const lnurlWithdrawRequest = (event, { service, amount, memo }) => dispatch => {
+  dispatch(setLnurlWithdrawParams({ amount, service, memo }))
 }
