+++ renderer/reducers/lnurl/selectors.js
@@ -1,4 +1,5 @@
 import { createSelector } from 'reselect'
+import { settingsSelectors } from 'reducers/settings'
 
 /**
  * @typedef {import('../index').State} State
@@ -13,12 +14,20 @@ import { createSelector } from 'reselect'
 const lnurlAuthParams = state => state.lnurl.lnurlAuthParams
 
 /**
- * lnurlAuthParams - Current lnurl channel params.
+ * lnurlChannelParams - Current lnurl channel params.
  *
  * @param {State} state Redux state
  * @returns {{service: string, secret: string}|null} Current lnurl auth paramaters
  */
 const lnurlChannelParams = state => state.lnurl.lnurlChannelParams
+
+/**
+ * lnurlWithdrawParams - Current lnurl channel params.
+ *
+ * @param {State} state Redux state
+ * @returns {{service: string, amount: string, memo: string}|null} Current lnurl auth paramaters
+ */
+const lnurlWithdrawParams = state => state.lnurl.lnurlWithdrawParams
 
 /**
  * willShowLnurlAuthPrompt - Boolean indicating wether lnurl auth prompt should show.
@@ -34,9 +43,23 @@ const willShowLnurlChannelPrompt = createSelector(lnurlChannelParams, params => 
   return Boolean(params)
 })
 
+/**
+ * willShowLnurlWithdrawPrompt - Boolean indicating wether lnurl withdraw prompt should show.
+ */
+export const willShowLnurlWithdrawPrompt = createSelector(
+  lnurlWithdrawParams,
+  settingsSelectors.currentConfig,
+  (params, config) => {
+    const promptEnabled = config.lnurl.requirePrompt
+    return Boolean(promptEnabled && params)
+  }
+)
+
 export default {
   lnurlAuthParams,
   lnurlChannelParams,
+  lnurlWithdrawParams,
   willShowLnurlAuthPrompt,
   willShowLnurlChannelPrompt,
+  willShowLnurlWithdrawPrompt,
 }
