+++ renderer/reducers/pay/constants.js
@@ -5,5 +5,3 @@ export const QUERY_ROUTES = 'QUERY_ROUTES'
 export const QUERY_ROUTES_SUCCESS = 'QUERY_ROUTES_SUCCESS'
 export const QUERY_ROUTES_FAILURE = 'QUERY_ROUTES_FAILURE'
 export const SET_REDIRECT_PAY_REQ = 'SET_REDIRECT_PAY_REQ'
-export const SET_LNURL_WITHDRAWAL_PARAMS = 'SET_LNURL_WITHDRAWAL_PARAMS'
-export const CLEAR_LNURL_WITHDRAWAL = 'CLEAR_LNURL_WITHDRAWAL'
