+++ renderer/reducers/pay/index.js
@@ -1,7 +1,6 @@
 import payReducer from './reducer'
 
 export default payReducer
-export paySelectors from './selectors'
 export * from './constants'
 export * from './reducer'
 export * from './ipc'
