--- renderer/reducers/pay/messages.js
@@ -1,10 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-// use pattern [reducer]_message
-
-/* eslint-disable max-len */
-export default defineMessages({
-  pay_lnurl_withdraw_started: 'Withdrawal requested. Please wait…',
-  pay_lnurl_withdraw_error: 'Unable to process withdrawal request from {service}: {reason}',
-  pay_lnurl_withdraw_success: 'Withdrawal request from {service} complete',
-})
