+++ renderer/reducers/pay/reducer.js
@@ -1,17 +1,12 @@
 import get from 'lodash/get'
-import { send } from 'redux-electron-ipc'
 import { grpc } from 'workers'
-import { getIntl } from '@zap/i18n'
 import createReducer from '@zap/utils/createReducer'
 import { estimateFeeRange } from '@zap/utils/fee'
 import { isPubkey, getTag } from '@zap/utils/crypto'
 import { settingsSelectors } from 'reducers/settings'
-import { showWarning } from 'reducers/notification'
 import { infoSelectors } from 'reducers/info'
 import { prepareKeysendProbe, prepareBolt11Probe } from 'reducers/payment'
-import { createInvoice } from 'reducers/invoice'
 import * as constants from './constants'
-import messages from './messages'
 
 const {
   QUERY_FEES,
@@ -21,8 +16,6 @@ const {
   QUERY_ROUTES_SUCCESS,
   QUERY_ROUTES_FAILURE,
   SET_REDIRECT_PAY_REQ,
-  SET_LNURL_WITHDRAWAL_PARAMS,
-  CLEAR_LNURL_WITHDRAWAL,
 } = constants
 
 // ------------------------------------
@@ -40,32 +33,12 @@ const initialState = {
   queryFeesError: null,
   queryRoutesError: null,
   redirectPayReq: null,
-  lnurlWithdrawParams: null,
   routes: [],
 }
 
 // ------------------------------------
 // Actions
 // ------------------------------------
-
-/**
- * clearLnurlWithdraw - Clears lnurl withdraw state
- *
- * @returns {(dispatch:Function) => void} Thunk
- */
-export const clearLnurlWithdraw = () => dispatch => {
-  dispatch({ type: CLEAR_LNURL_WITHDRAWAL })
-}
-
-/**
- * declineLnurlWithdraw - Cancels lnurl withdraw and clears params cache.
- *
- * @returns {(dispatch:Function) => void} Thunk
- */
-export const declineLnurlWithdraw = () => dispatch => {
-  dispatch(send('lnurlCancelWithdraw'))
-  dispatch(clearLnurlWithdraw())
-}
 
 /**
  * setRedirectPayReq - Set payment request to initiate payment flow to a specific address / payment request.
@@ -77,54 +50,6 @@ export function setRedirectPayReq(redirectPayReq) {
   return {
     type: SET_REDIRECT_PAY_REQ,
     redirectPayReq,
-  }
-}
-
-/**
- * setLnurlWithdrawParams - Set request details.
- *
- * @param {object|null} params lnurl request details or null to clear
- * @returns {object} Action
- */
-export function setLnurlWithdrawParams(params) {
-  return {
-    type: SET_LNURL_WITHDRAWAL_PARAMS,
-    params,
-  }
-}
-
-/**
- * finishLnurlWithdraw - Concludes lnurl withdraw request processing by sending our ln PR to the service.
- *
- * @returns {(dispatch:Function, getState:Function) => Promise<void>} Thunk
- */
-export const finishLnurlWithdraw = () => async (dispatch, getState) => {
-  const state = getState()
-  if (state.pay.lnurlWithdrawParams) {
-    const { service, amount, memo } = getState().pay.lnurlWithdrawParams
-
-    // Show notification.
-    dispatch(
-      showWarning(getIntl().formatMessage(messages.pay_lnurl_withdraw_started), {
-        timeout: 3000,
-        isProcessing: true,
-        payload: { service },
-      })
-    )
-
-    // Create invoice.
-    const { paymentRequest } = await dispatch(
-      createInvoice({
-        amount,
-        memo,
-        cryptoUnit: 'msats',
-        isPrivate: true,
-      })
-    )
-
-    // Hand off to main lnurl service to complete the process.
-    dispatch(send('lnurlFinishWithdraw', { paymentRequest }))
-    dispatch(clearLnurlWithdraw())
   }
 }
 
@@ -261,12 +186,6 @@ const ACTION_HANDLERS = {
   },
   [SET_REDIRECT_PAY_REQ]: (state, { redirectPayReq }) => {
     state.redirectPayReq = redirectPayReq
-  },
-  [SET_LNURL_WITHDRAWAL_PARAMS]: (state, { params }) => {
-    state.lnurlWithdrawParams = params
-  },
-  [CLEAR_LNURL_WITHDRAWAL]: state => {
-    state.lnurlWithdrawParams = null
   },
 }
