+++ stories/dialogs/lnurl-withdrawal-prompt.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import { LnurlWithdrawPrompt } from 'components/Pay'
+import { LnurlWithdrawPrompt } from 'components/Lnurl'
 import { Window } from '../helpers'
 
 storiesOf('Dialogs', module)
