+++ test/unit/reducers/pay.spec.js
@@ -7,7 +7,6 @@ import reducer, {
   QUERY_ROUTES_SUCCESS,
   QUERY_ROUTES_FAILURE,
   SET_REDIRECT_PAY_REQ,
-  SET_LNURL_WITHDRAWAL_PARAMS,
 } from 'reducers/pay'
 
 describe('reducers', () => {
@@ -71,14 +70,6 @@ describe('reducers', () => {
       const action = {
         type: SET_REDIRECT_PAY_REQ,
         redirectPayReq: 'lightning:abc123',
-      }
-      expect(snapshotDiff(reducer(undefined, {}), reducer(undefined, action))).toMatchSnapshot()
-    })
-
-    it('should handle SET_LNURL_WITHDRAWAL_PARAMS', () => {
-      const action = {
-        type: SET_LNURL_WITHDRAWAL_PARAMS,
-        params: { memo: '1', amount: 100 },
       }
       expect(snapshotDiff(reducer(undefined, {}), reducer(undefined, action))).toMatchSnapshot()
     })
